+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/build.py
@@ -1,11 +1,9 @@
 from dagster import OpExecutionContext
-from dagster_dbt import DbtCli, DbtManifest, dbt_assets
+from dagster_dbt import DbtCli, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
-manifest = DbtManifest.read(path=MANIFEST_PATH)
 
-
-@dbt_assets(manifest=manifest)
+@dbt_assets(manifest=MANIFEST_PATH)
 def my_dbt_assets(context: OpExecutionContext, dbt: DbtCli):
     yield from dbt.cli(["build"], context=context).stream()
