+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/customize_asset_keys.py
@@ -1,20 +1,17 @@
 from typing import Any, Mapping
 
 from dagster import AssetKey, OpExecutionContext
-from dagster_dbt import DbtCli, DbtManifest, dbt_assets
+from dagster_dbt import DagsterDbtTranslator, DbtCli, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
 
-class CustomizedDbtManifest(DbtManifest):
+class CustomDagsterDbtTranslator(DagsterDbtTranslator):
     @classmethod
     def node_info_to_asset_key(cls, node_info: Mapping[str, Any]) -> AssetKey:
         return AssetKey(["prefix", node_info["alias"]])
 
 
-manifest = CustomizedDbtManifest.read(path=MANIFEST_PATH)
-
-
-@dbt_assets(manifest=manifest)
+@dbt_assets(manifest=MANIFEST_PATH, dagster_dbt_translator=CustomDagsterDbtTranslator())
 def my_dbt_assets(context: OpExecutionContext, dbt: DbtCli):
     yield from dbt.cli(["build"], context=context).stream()
