+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/customize_description.py
@@ -1,12 +1,12 @@
 from typing import Any, Mapping
 
 from dagster import OpExecutionContext
-from dagster_dbt import DbtCli, DbtManifest, dbt_assets
+from dagster_dbt import DagsterDbtTranslator, DbtCli, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
 
-class CustomizedDbtManifest(DbtManifest):
+class CustomDagsterDbtTranslator(DagsterDbtTranslator):
     @classmethod
     def node_info_to_description(cls, node_info: Mapping[str, Any]) -> str:
         description_sections = []
@@ -22,9 +22,6 @@ class CustomizedDbtManifest(DbtManifest):
         return "\n\n".join(description_sections)
 
 
-manifest = CustomizedDbtManifest.read(path=MANIFEST_PATH)
-
-
-@dbt_assets(manifest=manifest)
+@dbt_assets(manifest=MANIFEST_PATH, dagster_dbt_translator=CustomDagsterDbtTranslator())
 def my_dbt_assets(context: OpExecutionContext, dbt: DbtCli):
     yield from dbt.cli(["build"], context=context).stream()
