+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/partitions.py
@@ -1,22 +1,20 @@
 import json
 
 from dagster import DailyPartitionsDefinition, OpExecutionContext
-from dagster_dbt import DbtCli, DbtManifest, dbt_assets
+from dagster_dbt import DbtCli, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
 DBT_SELECT_SEED = "resource_type:seed"
 
-manifest = DbtManifest.read(path=MANIFEST_PATH)
 
-
-@dbt_assets(manifest=manifest, select=DBT_SELECT_SEED)
+@dbt_assets(manifest=MANIFEST_PATH, select=DBT_SELECT_SEED)
 def dbt_seed_assets(context: OpExecutionContext, dbt: DbtCli):
     yield from dbt.cli(["seed"], context=context).stream()
 
 
 @dbt_assets(
-    manifest=manifest,
+    manifest=MANIFEST_PATH,
     exclude=DBT_SELECT_SEED,
     partitions_def=DailyPartitionsDefinition(start_date="2023-05-01"),
 )
