+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/seed_run_test.py
@@ -1,12 +1,10 @@
 from dagster import OpExecutionContext
-from dagster_dbt import DbtCli, DbtManifest, dbt_assets
+from dagster_dbt import DbtCli, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
-manifest = DbtManifest.read(path=MANIFEST_PATH)
 
-
-@dbt_assets(manifest=manifest)
+@dbt_assets(manifest=MANIFEST_PATH)
 def my_dbt_assets(context: OpExecutionContext, dbt: DbtCli):
     dbt_commands = [
         ["seed"],
