+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/separate_asset_defs.py
@@ -1,17 +1,15 @@
 from dagster import OpExecutionContext
-from dagster_dbt import DbtCli, DbtManifest, dbt_assets
+from dagster_dbt import DbtCli, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
-manifest = DbtManifest.read(path=MANIFEST_PATH)
 
-
-@dbt_assets(manifest=manifest, select="resource_type:seed")
+@dbt_assets(manifest=MANIFEST_PATH, select="resource_type:seed")
 def dbt_seed_assets(context: OpExecutionContext, dbt: DbtCli):
     yield from dbt.cli(["seed"], context=context).stream()
 
 
-@dbt_assets(manifest=manifest, select="fqn:staging.*")
+@dbt_assets(manifest=MANIFEST_PATH, select="fqn:staging.*")
 def dbt_staging_assets(context: OpExecutionContext, dbt: DbtCli):
     dbt_commands = [
         ["run"],
@@ -22,7 +20,7 @@ def dbt_staging_assets(context: OpExecutionContext, dbt: DbtCli):
         yield from dbt.cli(dbt_command, context=context).stream()
 
 
-@dbt_assets(manifest=manifest, select="fqn:customers fqn:orders")
+@dbt_assets(manifest=MANIFEST_PATH, select="fqn:customers fqn:orders")
 def my_dbt_assets(context: OpExecutionContext, dbt: DbtCli):
     dbt_commands = [
         ["run"],
