+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/constants.py
@@ -1,4 +1,6 @@
+from pathlib import Path
+
 from dagster import file_relative_path
 
 DBT_PROJECT_DIR = file_relative_path(__file__, "../jaffle_shop")
-MANIFEST_PATH = file_relative_path(__file__, "../jaffle_shop/target/manifest.json")
+MANIFEST_PATH = Path(__file__) / "jaffle_shop" / "target" / "manifest.json"
