+++ python_modules/libraries/dagster-dbt/dagster_dbt/asset_utils.py
@@ -27,8 +27,6 @@ from dagster._utils.merger import merge_dicts
 
 from .utils import input_name_fn, output_name_fn
 
-MANIFEST_METADATA_KEY = "dagster_dbt/manifest"
-
 ###################
 # DEFAULT FUNCTIONS
 ###################
