+++ python_modules/libraries/dagster-dbt/dagster_dbt/core/__init__.py
@@ -7,7 +7,5 @@ from .resources_v2 import (
     DbtCli as DbtCli,
     DbtCliEventMessage as DbtCliEventMessage,
     DbtCliInvocation as DbtCliInvocation,
-    DbtManifest as DbtManifest,
-    DbtManifestAssetSelection as DbtManifestAssetSelection,
 )
 from .types import DbtCliOutput as DbtCliOutput
