+++ python_modules/libraries/dagster-dbt/dagster_dbt/utils.py
@@ -320,3 +320,16 @@ def select_unique_ids_from_manifest(
     selector = graph_selector.NodeSelector(graph, manifest, previous_state=previous_state)
     selected, _ = selector.select_nodes(parsed_spec)
     return selected
+
+
+def get_node_info_by_dbt_unique_id_from_manifest(
+    manifest: Mapping[str, Any]
+) -> Mapping[str, Mapping[str, Any]]:
+    """A mapping of a dbt node's unique id to the node's dictionary representation in the manifest.
+    """
+    return {
+        **manifest["nodes"],
+        **manifest["sources"],
+        **manifest["exposures"],
+        **manifest["metrics"],
+    }
